+++ category/views.py
@@ -7,6 +7,8 @@ from .serializers import CategorySerializer, CategoryPostSerializer
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly 
+from .permissions import CategoryPermission
 
 class CategoryListView(View):
     def get(self, request):
@@ -60,6 +62,7 @@ class DeleteCategoryView(LoginRequiredMixin, PermissionRequiredMixin, View):
 
 ### API ###
 class CategoryListAPI(APIView):
+    permission_classes = [IsAuthenticatedOrReadOnly, CategoryPermission]
     def get(self, request):
         categories = Category.objects.all()
         serializer = CategorySerializer(instance=categories, many=True)
@@ -73,13 +76,16 @@ class CategoryListAPI(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 class CategoryDetailAPI(APIView):
+    permission_classes = [IsAuthenticated, CategoryPermission]
     def get(self, request, category_id):
         category = get_object_or_404(Category, pk=category_id)
+        self.check_object_permissions(request, category)
         serializer = CategorySerializer(instance=category)
         return Response(serializer.data)
 
     def put(self, request, category_id):
         category = get_object_or_404(Category, pk=category_id)
+        self.check_object_permissions(request, category)
         serializer = CategoryPostSerializer(data=request.data, instance=category)
         if serializer.is_valid():
             serializer.save()
@@ -88,5 +94,6 @@ class CategoryDetailAPI(APIView):
 
     def delete(self, request, category_id):
         category = get_object_or_404(Category, pk=category_id)
+        self.check_object_permissions(request, category)
         category.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
